@@ -60,61 +60,61 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>21</maven.compiler.source>
-        <maven.compiler.target>21</maven.compiler.target>
+        <maven.compiler.source>25</maven.compiler.source>
+        <maven.compiler.target>25</maven.compiler.target>
 
         <!-- BOM Versions (for reference - not used as BOMs, using explicit versions instead) -->
-        <spring-boot.version>3.4.3</spring-boot.version>
+        <spring-boot.version>3.5.10</spring-boot.version>
         <jakarta-ee.version>10.0.0</jakarta-ee.version>
-        <opentelemetry.version>1.45.0</opentelemetry.version>
-        <opentelemetry-instrumentation.version>2.10.0</opentelemetry-instrumentation.version>
+        <opentelemetry.version>1.52.0</opentelemetry.version>
+        <opentelemetry-instrumentation.version>2.18.1</opentelemetry-instrumentation.version>
         <resilience4j.version>2.3.0</resilience4j.version>
-        <testcontainers.version>1.20.4</testcontainers.version>
+        <testcontainers.version>1.21.3</testcontainers.version>
         <jmh.version>1.37</jmh.version>
-        <jacoco.version>0.8.12</jacoco.version>
+        <jacoco.version>0.8.13</jacoco.version>
 
         <!-- Jakarta EE & Framework -->
         <jakarta.servlet.version>6.1.0</jakarta.servlet.version>
         <jakarta.annotation.version>3.0.0</jakarta.annotation.version>
-        <jakarta.persistence.version>3.1.0</jakarta.persistence.version>
-        <jakarta.xml.bind.version>3.0.1</jakarta.xml.bind.version>
-        <jakarta.mail.version>2.1.3</jakarta.mail.version>
-        <angus.mail.version>2.0.3</angus.mail.version>
-        <jakarta.activation.version>2.1.0</jakarta.activation.version>
-        <angus.activation.version>2.0.2</angus.activation.version>
+        <jakarta.persistence.version>3.2.0</jakarta.persistence.version>
+        <jakarta.xml.bind.version>4.0.5</jakarta.xml.bind.version>
+        <jakarta.mail.version>2.1.5</jakarta.mail.version>
+        <angus.mail.version>2.0.5</angus.mail.version>
+        <jakarta.activation.version>2.1.4</jakarta.activation.version>
+        <angus.activation.version>2.0.3</angus.activation.version>
         <jakarta.faces.version>4.1.2</jakarta.faces.version>
-        <jakarta.cdi.version>4.0.1</jakarta.cdi.version>
+        <jakarta.cdi.version>4.1.0</jakarta.cdi.version>
 
         <!-- ORM -->
-        <hibernate.version>6.6.5.Final</hibernate.version>
+        <hibernate.version>6.6.42.Final</hibernate.version>
 
         <!-- Core Libraries -->
-        <commons.lang3.version>3.17.0</commons.lang3.version>
-        <commons.io.version>2.18.0</commons.io.version>
-        <commons.codec.version>1.17.1</commons.codec.version>
-        <commons.vfs2.version>2.9.0</commons.vfs2.version>
-        <commons.collections4.version>4.4</commons.collections4.version>
-        <commons.dbcp2.version>2.12.0</commons.dbcp2.version>
-        <commons.fileupload.version>1.5</commons.fileupload.version>
-        <commons.pool2.version>2.12.0</commons.pool2.version>
-        <commons.text.version>1.13.0</commons.text.version>
+        <commons.lang3.version>3.20.0</commons.lang3.version>
+        <commons.io.version>2.21.0</commons.io.version>
+        <commons.codec.version>1.19.0</commons.codec.version>
+        <commons.vfs2.version>2.10.0</commons.vfs2.version>
+        <commons.collections4.version>4.5.0-M2</commons.collections4.version>
+        <commons.dbcp2.version>2.14.0</commons.dbcp2.version>
+        <commons.fileupload.version>1.6.0</commons.fileupload.version>
+        <commons.pool2.version>2.13.1</commons.pool2.version>
+        <commons.text.version>1.15.0</commons.text.version>
 
         <!-- Database Drivers -->
-        <h2.version>2.3.232</h2.version>
-        <postgresql.version>42.7.4</postgresql.version>
-        <mysql.version>9.1.0</mysql.version>
+        <h2.version>2.4.240</h2.version>
+        <postgresql.version>42.7.7</postgresql.version>
+        <mysql.version>9.4.0</mysql.version>
         <derby.version>10.17.1.0</derby.version>
         <hsqldb.version>2.7.4</hsqldb.version>
-        <hikaricp.version>6.2.1</hikaricp.version>
+        <hikaricp.version>7.0.2</hikaricp.version>
 
-        <!-- Logging - updated to 2.25.3 (bug fixes, GraalVM support) -->
+        <!-- Logging - updated to 3.0.0-beta3 (log4j2 v3 beta) -->
         <log4j.version>2.25.3</log4j.version>
         <slf4j.version>2.0.17</slf4j.version>
         <jboss.logging.version>3.6.1.Final</jboss.logging.version>
 
         <!-- JSON Processing -->
-        <jackson.version>2.18.3</jackson.version>
-        <gson.version>2.11.0</gson.version>
+        <jackson.version>2.19.4</jackson.version>
+        <gson.version>2.13.2</gson.version>
 
         <!-- XML Processing -->
         <jdom.version>2.0.6.1</jdom.version>
@@ -1373,10 +1373,11 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.14.0</version>
                 <configuration>
-                    <source>21</source>
-                    <target>21</target>
-                    <release>21</release>
+                    <source>25</source>
+                    <target>25</target>
+                    <release>25</release>
                     <compilerArgs>
+                        <arg>--enable-preview</arg>
                         <arg>-Xlint:all</arg>
                     </compilerArgs>
                 </configuration>
@@ -1522,7 +1520,7 @@
                         <groupId>org.apache.maven.plugins</groupId>
                         <artifactId>maven-compiler-plugin</artifactId>
                         <configuration>
-                            <release>21</release>
+                            <release>25</release>
                             <compilerArgs>
                                 <arg>--enable-preview</arg>
                                 <arg>-Xlint:all</arg>
