@@ -24,10 +24,10 @@ import org.w3c.dom.ls.LSResourceResolver;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.UnsupportedEncodingException;
 import java.net.MalformedURLException;
 import java.net.URI;
 import java.net.URL;
+import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Scanner;
@@ -76,8 +76,8 @@ public class ResourceResolver implements LSResourceResolver {
     }
 
 
-    private InputStream stringToStream(String s) throws UnsupportedEncodingException {
-        return new ByteArrayInputStream(s.getBytes("UTF-8"));
+    private InputStream stringToStream(String s) {
+        return new ByteArrayInputStream(s.getBytes(StandardCharsets.UTF_8));
     }
 
 }
