@@ -1,9 +1,11 @@
 package org.yawlfoundation.yawl.patternmatching;
 
-import junit.framework.TestCase;
+import org.junit.jupiter.api.Test;
 import org.yawlfoundation.yawl.engine.YWorkItem;
 import org.yawlfoundation.yawl.elements.YTimerParameters;
 import org.yawlfoundation.yawl.schema.YSchemaVersion;
+
+import static org.junit.jupiter.api.Assertions.*;
 
 /**
  * Tests for enum exhaustiveness in switch expressions
@@ -18,23 +20,25 @@ import org.yawlfoundation.yawl.schema.YSchemaVersion;
  * Author: YAWL Foundation
  * Date: 2026-02-16
  */
-public class EnumExhaustivenessTest extends TestCase {
+class EnumExhaustivenessTest {
 
     // Test YWorkItem.Completion enum exhaustiveness
-    public void testCompletionEnum_AllValuesHandled() {
+    @Test
+    void testCompletionEnum_AllValuesHandled() {
         for (YWorkItem.Completion completion : YWorkItem.Completion.values()) {
             // Each value should produce a valid status
             String status = getCompletionStatusName(completion);
-            assertNotNull("Status should not be null for " + completion, status);
-            assertFalse("Status should not be empty for " + completion,
-                       status.isEmpty());
+            assertNotNull(status, "Status should not be null for " + completion);
+            assertFalse(status.isEmpty(),
+                       "Status should not be empty for " + completion);
         }
     }
 
-    public void testCompletionEnum_NoMissingCases() {
+    @Test
+    void testCompletionEnum_NoMissingCases() {
         // Verify count matches expected
-        assertEquals("Should have exactly 3 completion types",
-                    3, YWorkItem.Completion.values().length);
+        assertEquals(3, YWorkItem.Completion.values().length,
+                    "Should have exactly 3 completion types");
 
         // Verify all expected values exist
         assertNotNull(YWorkItem.Completion.Normal);
@@ -43,20 +47,22 @@ public class EnumExhaustivenessTest extends TestCase {
     }
 
     // Test YTimerParameters.TimerType enum exhaustiveness
-    public void testTimerTypeEnum_AllValuesHandled() {
+    @Test
+    void testTimerTypeEnum_AllValuesHandled() {
         for (YTimerParameters.TimerType timerType : YTimerParameters.TimerType.values()) {
             // Each value should produce a valid description
             String desc = getTimerTypeDescription(timerType);
-            assertNotNull("Description should not be null for " + timerType, desc);
-            assertFalse("Description should not be empty for " + timerType,
-                       desc.isEmpty());
+            assertNotNull(desc, "Description should not be null for " + timerType);
+            assertFalse(desc.isEmpty(),
+                       "Description should not be empty for " + timerType);
         }
     }
 
-    public void testTimerTypeEnum_NoMissingCases() {
+    @Test
+    void testTimerTypeEnum_NoMissingCases() {
         // Verify count matches expected
-        assertEquals("Should have exactly 3 timer types",
-                    3, YTimerParameters.TimerType.values().length);
+        assertEquals(3, YTimerParameters.TimerType.values().length,
+                    "Should have exactly 3 timer types");
 
         // Verify all expected values exist
         assertNotNull(YTimerParameters.TimerType.Expiry);
@@ -65,20 +71,22 @@ public class EnumExhaustivenessTest extends TestCase {
     }
 
     // Test YTimerParameters.TriggerType enum exhaustiveness
-    public void testTriggerTypeEnum_AllValuesHandled() {
+    @Test
+    void testTriggerTypeEnum_AllValuesHandled() {
         for (YTimerParameters.TriggerType trigger : YTimerParameters.TriggerType.values()) {
             // Each value should produce a valid status
             String status = getTriggerStatusName(trigger);
-            assertNotNull("Status should not be null for " + trigger, status);
-            assertFalse("Status should not be empty for " + trigger,
-                       status.isEmpty());
+            assertNotNull(status, "Status should not be null for " + trigger);
+            assertFalse(status.isEmpty(),
+                       "Status should not be empty for " + trigger);
         }
     }
 
-    public void testTriggerTypeEnum_NoMissingCases() {
+    @Test
+    void testTriggerTypeEnum_NoMissingCases() {
         // Verify count matches expected
-        assertEquals("Should have exactly 2 trigger types",
-                    2, YTimerParameters.TriggerType.values().length);
+        assertEquals(2, YTimerParameters.TriggerType.values().length,
+                    "Should have exactly 2 trigger types");
 
         // Verify all expected values exist
         assertNotNull(YTimerParameters.TriggerType.OnEnabled);
@@ -86,20 +94,22 @@ public class EnumExhaustivenessTest extends TestCase {
     }
 
     // Test YSchemaVersion enum exhaustiveness
-    public void testSchemaVersionEnum_AllValuesHandled() {
+    @Test
+    void testSchemaVersionEnum_AllValuesHandled() {
         for (YSchemaVersion version : YSchemaVersion.values()) {
             // Each value should produce a valid beta flag
             boolean isBeta = version.isBetaVersion();
             // Should not throw exception - verify by asserting non-null Boolean
-            assertNotNull("Beta flag should be determinable for " + version,
-                         Boolean.valueOf(isBeta));
+            assertNotNull(Boolean.valueOf(isBeta),
+                         "Beta flag should be determinable for " + version);
         }
     }
 
-    public void testSchemaVersionEnum_NoMissingCases() {
+    @Test
+    void testSchemaVersionEnum_NoMissingCases() {
         // Verify count matches expected
-        assertEquals("Should have exactly 10 schema versions",
-                    10, YSchemaVersion.values().length);
+        assertEquals(10, YSchemaVersion.values().length,
+                    "Should have exactly 10 schema versions");
 
         // Verify all expected values exist
         assertNotNull(YSchemaVersion.Beta2);
@@ -115,7 +125,8 @@ public class EnumExhaustivenessTest extends TestCase {
     }
 
     // Test that adding new enum values would fail these tests
-    public void testEnumCounts_NoUnexpectedValues() {
+    @Test
+    void testEnumCounts_NoUnexpectedValues() {
         // If someone adds a new enum value, these tests should catch it
         int completionCount = 0;
         int timerTypeCount = 0;
@@ -135,47 +146,53 @@ public class EnumExhaustivenessTest extends TestCase {
             schemaVersionCount++;
         }
 
-        assertEquals("Completion enum count mismatch", 3, completionCount);
-        assertEquals("Timer type enum count mismatch", 3, timerTypeCount);
-        assertEquals("Trigger type enum count mismatch", 2, triggerTypeCount);
-        assertEquals("Schema version enum count mismatch", 10, schemaVersionCount);
+        assertEquals(3, completionCount, "Completion enum count mismatch");
+        assertEquals(3, timerTypeCount, "Timer type enum count mismatch");
+        assertEquals(2, triggerTypeCount, "Trigger type enum count mismatch");
+        assertEquals(10, schemaVersionCount, "Schema version enum count mismatch");
     }
 
     // Test enum ordinal consistency
-    public void testEnumOrdinals_Sequential() {
+    @Test
+    void testEnumOrdinals_Sequential() {
         // Completion ordinals should be 0, 1, 2
         YWorkItem.Completion[] completions = YWorkItem.Completion.values();
         for (int i = 0; i < completions.length; i++) {
-            assertEquals("Completion ordinal should match index", i, completions[i].ordinal());
+            assertEquals(i, completions[i].ordinal(),
+                        "Completion ordinal should match index");
         }
 
         // Timer type ordinals should be 0, 1, 2
         YTimerParameters.TimerType[] timerTypes = YTimerParameters.TimerType.values();
         for (int i = 0; i < timerTypes.length; i++) {
-            assertEquals("Timer type ordinal should match index", i, timerTypes[i].ordinal());
+            assertEquals(i, timerTypes[i].ordinal(),
+                        "Timer type ordinal should match index");
         }
 
         // Trigger type ordinals should be 0, 1
         YTimerParameters.TriggerType[] triggers = YTimerParameters.TriggerType.values();
         for (int i = 0; i < triggers.length; i++) {
-            assertEquals("Trigger type ordinal should match index", i, triggers[i].ordinal());
+            assertEquals(i, triggers[i].ordinal(),
+                        "Trigger type ordinal should match index");
         }
 
         // Schema version ordinals should be 0-9
         YSchemaVersion[] versions = YSchemaVersion.values();
         for (int i = 0; i < versions.length; i++) {
-            assertEquals("Schema version ordinal should match index", i, versions[i].ordinal());
+            assertEquals(i, versions[i].ordinal(),
+                        "Schema version ordinal should match index");
         }
     }
 
     // Test that all enum values are distinct
-    public void testEnumValues_AllDistinct() {
+    @Test
+    void testEnumValues_AllDistinct() {
         // Test completion enum
         YWorkItem.Completion[] completions = YWorkItem.Completion.values();
         for (int i = 0; i < completions.length; i++) {
             for (int j = i + 1; j < completions.length; j++) {
-                assertFalse("Completion values should be distinct",
-                           completions[i].equals(completions[j]));
+                assertFalse(completions[i].equals(completions[j]),
+                           "Completion values should be distinct");
             }
         }
 
@@ -183,8 +200,8 @@ public class EnumExhaustivenessTest extends TestCase {
         YTimerParameters.TimerType[] timerTypes = YTimerParameters.TimerType.values();
         for (int i = 0; i < timerTypes.length; i++) {
             for (int j = i + 1; j < timerTypes.length; j++) {
-                assertFalse("Timer type values should be distinct",
-                           timerTypes[i].equals(timerTypes[j]));
+                assertFalse(timerTypes[i].equals(timerTypes[j]),
+                           "Timer type values should be distinct");
             }
         }
 
@@ -192,8 +209,8 @@ public class EnumExhaustivenessTest extends TestCase {
         YTimerParameters.TriggerType[] triggers = YTimerParameters.TriggerType.values();
         for (int i = 0; i < triggers.length; i++) {
             for (int j = i + 1; j < triggers.length; j++) {
-                assertFalse("Trigger type values should be distinct",
-                           triggers[i].equals(triggers[j]));
+                assertFalse(triggers[i].equals(triggers[j]),
+                           "Trigger type values should be distinct");
             }
         }
 
@@ -201,8 +218,8 @@ public class EnumExhaustivenessTest extends TestCase {
         YSchemaVersion[] versions = YSchemaVersion.values();
         for (int i = 0; i < versions.length; i++) {
             for (int j = i + 1; j < versions.length; j++) {
-                assertFalse("Schema version values should be distinct",
-                           versions[i].equals(versions[j]));
+                assertFalse(versions[i].equals(versions[j]),
+                           "Schema version values should be distinct");
             }
         }
     }
